Recognize WideInt as int in C++ leaf collector
C++ ``_collect_int_leaves`` only collected values when the inferred
element type was ``int``; after the global type-inference change
returns ``WideInt`` for mixed-magnitude int collections, the C++
narrowness resolver got an empty leaf list and fell back to ``int``,
re-introducing the narrowing error the WideInt path was meant to fix.
Treat ``WideInt`` as an int leaf so the resolver picks ``long long``
and the emitted ``std::initializer_list<long long>`` compiles.

+++ tests/integration/cases/set_mixed_int_widths/Cpp_combined.cpp
@@ -1,11 +1,11 @@
 #include <initializer_list>
 int main() {
-auto my_data = std::initializer_list<int>{
+auto my_data = std::initializer_list<long long>{
     1,
     1099511627776,
 };
 (void)my_data;
-my_data = std::initializer_list<int>{
+my_data = std::initializer_list<long long>{
     1,
     1099511627776,
 };
