+++ tests/integration/cases/set_mixed_int_widths/Cpp_type_hints_safe.cpp
@@ -1,6 +1,6 @@
 #include <initializer_list>
 int main() {
-auto my_data = std::initializer_list<int>{
+auto my_data = std::initializer_list<long long>{
     1,
     1099511627776,
 };
